+++ tools/build.py
@@ -13,7 +13,7 @@ import sys
 sys.path.insert(0, f'{os.getcwd()}/../../novelibre/tools')
 from package_builder import PackageBuilder
 
-VERSION = '5.8.0'
+VERSION = '5.9.0'
 
 
 class PluginBuilder(PackageBuilder):